chore(kern): move task & scheduler to own mods
